FletApp/myapp/requirements.txt
flet==0.22.*
openpyxl==3.1.5
pyserial==3.5
